+++ changelog.md
@@ -1,3 +1,9 @@
+1.5.5
+------
+- Updated deprecated functions
+- CO-360 - Added missing translations 
+- CO-358 - Added an feature to delete inconsistent specific data through the module configuration 
+
 1.5.4
 ------
 - Added Connector Plugin support